+++ benches/dijkstra_weighted/small_graph1.rs
@@ -47,7 +47,7 @@ fn vec_vec(bencher: Bencher) {
     let mut heap = BinaryHeap::from(HEAP);
 
     bencher.bench_local(|| {
-        graph.dijkstra(|acc, w| acc + w, &mut dist, &mut heap);
+        graph.min_distances(|acc, w| acc + w, &mut dist, &mut heap);
 
         dist
     });
@@ -71,7 +71,7 @@ fn vec_hash_set(bencher: Bencher) {
     let mut heap = BinaryHeap::from(HEAP);
 
     bencher.bench_local(|| {
-        graph.dijkstra(|acc, w| acc + w, &mut dist, &mut heap);
+        graph.min_distances(|acc, w| acc + w, &mut dist, &mut heap);
 
         dist
     });
@@ -95,7 +95,7 @@ fn arr_vec(bencher: Bencher) {
     let mut heap = BinaryHeap::from(HEAP);
 
     bencher.bench_local(|| {
-        graph.dijkstra(|acc, w| acc + w, &mut dist, &mut heap);
+        graph.min_distances(|acc, w| acc + w, &mut dist, &mut heap);
 
         dist
     });
@@ -119,7 +119,7 @@ fn arr_hash_set(bencher: Bencher) {
     let mut heap = BinaryHeap::from(HEAP);
 
     bencher.bench_local(|| {
-        graph.dijkstra(|acc, w| acc + w, &mut dist, &mut heap);
+        graph.min_distances(|acc, w| acc + w, &mut dist, &mut heap);
 
         dist
     });
@@ -143,7 +143,7 @@ fn hash_map_vec(bencher: Bencher) {
     let mut heap = BinaryHeap::from(HEAP);
 
     bencher.bench_local(|| {
-        graph.dijkstra(|acc, w| acc + w, &mut dist, &mut heap);
+        graph.min_distances(|acc, w| acc + w, &mut dist, &mut heap);
 
         dist
     });
@@ -167,7 +167,7 @@ fn hash_map_hash_set(bencher: Bencher) {
     let mut heap = BinaryHeap::from(HEAP);
 
     bencher.bench_local(|| {
-        graph.dijkstra(|acc, w| acc + w, &mut dist, &mut heap);
+        graph.min_distances(|acc, w| acc + w, &mut dist, &mut heap);
 
         dist
     });
@@ -191,7 +191,7 @@ fn hash_map_hash_map(bencher: Bencher) {
     let mut heap = BinaryHeap::from(HEAP);
 
     bencher.bench_local(|| {
-        graph.dijkstra(|acc, w| acc + w, &mut dist, &mut heap);
+        graph.min_distances(|acc, w| acc + w, &mut dist, &mut heap);
 
         dist
     });