+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Changed
 
-- Test `DijkstraWeighted.dijkstra_sssp_weighted` for every source vertex.
+- Test `DijkstraWeighted.dijkstra_sssp_unweighted` for every source vertex.
 - Make `CHANGELOG.md` adhere to [keep a changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.2.2] - 2024-03-31