+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Add doctest example for `dijkstra_sssp_weighted`.
+
 ## [0.2.3] - 2024-03-31
 
 ### Added
@@ -9,7 +15,7 @@
 
 ### Changed
 
-- Test `DijkstraWeighted.dijkstra_sssp_unweighted` for every source vertex.
+- Test `DijkstraUnweighted.dijkstra_sssp_unweighted` for every source vertex.
 - Make `CHANGELOG.md` adhere to [keep a changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.2.2] - 2024-03-31